Dynamically choose which player type to use
This saves you from having to edit the source code to go between playing
the AI yourself and letting the AI play itself.

I also had to change the `Player` trait to be object-safe, but the new
version is closer to the UCI protocol anyways, so this is a net
positive.